lpanda: Local Political Actor Network Diachronic Analysis Tools

Provides functions to prepare, visualize, and analyse diachronic network data on local political actors, with a particular focus on the development of local party systems and identification of actor groups. Formalizes and automates a continuity diagram method that has been previously applied in research on Czech local politics, e.g. Bubenicek and Kubalek (2010, ISSN:1803-8220), Kubalek and Bubenicek (2012, ISSN:1803-8220), and Cmejrek, Bubenicek, and Copik (2010, ISBN:978-80-247-3061-5). The package also includes several example datasets derived from Czech municipal elections, compiled from official election results, field research, and previously published case studies on Czech local politics.

See also

Author

Maintainer: Vaclav Bubenicek bubenicek@pef.czu.cz (ORCID) [copyright holder]